# Parallelize encryption to improve performance