@Shopify/locutus 

Turns out encryption is pretty slow.

![yak](https://s-media-cache-ak0.pinimg.com/236x/21/89/57/218957c9f2d17cfbc8e029a5f1221e0d.jpg)

Time to decrypt test file fetched from  https://edg.epa.gov/data.json and encrypted with a random key:

On my macbook:

* Go 1.4, GOMAXPROCS=1, master branch: 7.8s
* Go 1.5, GOMAXPROCS=1, master branch: 5.8s
* Go 1.5, GOMAXPROCS=1, this branch: 6.9s
* Go 1.5, GOMAXPROCS=8, this branch: 1.9s

On production infrastructure (first is Go 1.4, 1 proc, master; second is Go 1.5, 32 procs, optimize branch):

```
burke@borg3.ash:~ $ ./test_deployed
7.371902681s
burke@borg3.ash:~ $ ./test_go_1.5_optimize_branch
794.795795ms
```