@burke 2015-08-27
Damn, turns out the profile result I was originally looking at that motivated this was from when we were still invoking ejson as `bundle exec ejson`. Our ejson files aren't actually big enough that this really pays off; most of the cost was from loading bundler and ruby. This'll only save us like 50-100ms.